Moving dochelpers out of the ARMI namespace
