+++ doc/conf.py
@@ -26,21 +26,27 @@ serve to show the default.
 """
 # ruff: noqa: E402
 import datetime
+import inspect
 import os
 import pathlib
 import re
 import shutil
+import subprocess
 import sys
 import warnings
 
-import sphinx_rtd_theme
+from docutils import nodes, statemachine
+from docutils.parsers.rst import Directive, directives
 from sphinx.domains.python import PythonDomain
+import sphinx_rtd_theme
 
 # handle python import locations for this execution
 PYTHONPATH = os.path.abspath("..")
 sys.path.insert(0, PYTHONPATH)
 # Also add to os.environ which will be used by the nbsphinx extension environment
 os.environ["PYTHONPATH"] = PYTHONPATH
+# Add dochelpers.py from doc/.static/ directory
+sys.path.insert(0, ".static")
 
 from armi import apps
 from armi import configure as armi_configure
@@ -48,7 +54,6 @@ from armi import context
 from armi import disableFutureConfigures
 from armi import meta
 from armi.bookkeeping import tests as bookkeepingTests
-from armi.utils.dochelpers import ExecDirective, PyReverse
 
 context.Mode.setMode(context.Mode.BATCH)
 
@@ -72,6 +77,141 @@ class PatchedPythonDomain(PythonDomain):
         return super(PatchedPythonDomain, self).resolve_xref(
             env, fromdocname, builder, typ, target, node, contnode
         )
+
+
+class ExecDirective(Directive):
+    """
+    Execute the specified python code and insert the output into the document.
+
+    The code is used as the body of a method, and must return a ``str``. The string result is
+    interpreted as reStructuredText.
+
+    Error handling informed by https://docutils.sourceforge.io/docs/howto/rst-directives.html#error-handling
+    The self.error function should both inform the documentation builder of the error and also
+    insert an error into the built documentation.
+
+    Warning
+    -------
+    This only works on a single node in the doctree, so the rendered code may not contain any new
+    section names or labels. They will result in ``WARNING: Unexpected section title`` warnings.
+    """
+
+    has_content = True
+
+    def run(self):
+        try:
+            code = inspect.cleandoc(
+                """
+            def usermethod():
+                {}
+            """
+            ).format("\n    ".join(self.content))
+            exec(code)
+            result = locals()["usermethod"]()
+
+            if result is None:
+
+                raise self.error(
+                    "Return value needed! The body of your `.. exec::` is used as a "
+                    "function call that must return a value."
+                )
+
+            para = nodes.container()
+            lines = statemachine.StringList(result.split("\n"))
+            self.state.nested_parse(lines, self.content_offset, para)
+            return [para]
+        except Exception as e:
+            docname = self.state.document.settings.env.docname
+            raise self.error(
+                "Unable to execute embedded doc code at {}:{} ... {}\n{}".format(
+                    docname, self.lineno, datetime.datetime.now(), str(e)
+                )
+            )
+
+
+class PyReverse(Directive):
+    """Runs pyreverse to generate UML for specified module name and options.
+
+    The directive accepts the same arguments as pyreverse, except you should not specify
+    ``--project`` or ``-o`` (output format). These are automatically specified.
+
+    If you pass ``-c`` to this, the figure generated is forced to be the className.png
+    like ``BurnMatrix.png``. For .gitignore purposes, this is a pain. Thus, we
+    auto-prefix ALL images generated by this directive with ``pyrev_``.
+    """
+
+    has_content = True
+    required_arguments = 1
+    optional_arguments = 50
+    option_spec = {
+        "alt": directives.unchanged,
+        "height": directives.length_or_percentage_or_unitless,
+        "width": directives.length_or_percentage_or_unitless,
+        "align": lambda arg: directives.choice(arg, ("left", "right", "center")),
+        "filename": directives.unchanged,
+    }
+
+    def run(self):
+        try:
+            args = list(self.arguments)
+            args.append("--project")
+            args.append(f"{args[0]}")
+            args.append("-opng")
+
+            # NOTE: cannot use "pylint.pyreverse.main.Run" because it calls `sys.exit`.
+            fig_name = self.options.get("filename", "classes_{}.png".format(args[0]))
+            command = [sys.executable, "-m", "pylint.pyreverse.main"]
+            print("Running {}".format(command + args))
+            env = dict(os.environ)
+            # apply any runtime path mods to the pythonpath env variable (e.g. sys.path
+            # mods made during doc confs)
+            env["PYTHONPATH"] = os.pathsep.join(sys.path)
+            subprocess.check_call(command + args, env=env)
+
+            try:
+                os.remove(os.path.join(APIDOC_REL, fig_name))
+            except OSError:
+                pass
+
+            shutil.move(fig_name, APIDOC_REL)
+            # add .gitignore helper prefix
+            shutil.move(
+                os.path.join(APIDOC_REL, fig_name),
+                os.path.join(APIDOC_REL, f"pyr_{fig_name}"),
+            )
+            new_content = [f".. figure:: /{APIDOC_REL}/pyr_{fig_name}"]
+
+            # assume we don't need the packages_, and delete.
+            try:
+                os.remove("packages_{}.png".format(args[0]))
+            except OSError:
+                pass
+
+            # pass the other args through (figure args like align)
+            for opt, val in self.options.items():
+                if opt in ("filename",):
+                    continue
+                new_content.append("    :{}: {}\n".format(opt, val))
+
+            new_content.append("\n")
+
+            for line in self.content:
+                new_content.append("    " + line)
+
+            para = nodes.container()
+            # tab_width = self.options.get('tab-width', self.state.document.settings.tab_width)
+            lines = statemachine.StringList(new_content)
+            self.state.nested_parse(lines, self.content_offset, para)
+            return [para]
+        except Exception as e:
+            docname = self.state.document.settings.env.docname
+            # add the error message directly to the built documentation and also tell the
+            # builder
+            raise self.error(
+                "Unable to execute embedded doc code at {}:{} ... {}\n{}".format(
+                    docname, self.lineno, datetime.datetime.now(), str(e)
+                )
+            )
 
 
 def autodoc_skip_member_handler(app, what, name, obj, skip, options):
