+++ doc/user/assembly_parameters_report.rst
@@ -9,7 +9,7 @@ This document lists all of the Assembly Parameters that are provided by the ARMI
 .. exec::
    from armi.reactor import assemblies
    from armi.reactor import assemblyParameters
-   from armi.utils.dochelpers import generateParamTable
+   from dochelpers import generateParamTable
 
    return generateParamTable(
        assemblies.Assembly, assemblyParameters.getAssemblyParameterDefinitions()
