+++ doc/user/block_parameters_report.rst
@@ -9,7 +9,7 @@ This document lists all of the Block Parameters that are provided by the ARMI Fr
 .. exec::
    from armi.reactor import blocks
    from armi.reactor import blockParameters
-   from armi.utils.dochelpers import generateParamTable
+   from dochelpers import generateParamTable
 
    return generateParamTable(
        blocks.Block, blockParameters.getBlockParameterDefinitions()
