+++ doc/user/reactor_parameters_report.rst
@@ -9,7 +9,7 @@ This document lists all of the Reactor Parameters that are provided by the ARMI 
 .. exec::
    from armi.reactor import reactors
    from armi.reactor import reactorParameters
-   from armi.utils.dochelpers import generateParamTable
+   from dochelpers import generateParamTable
 
    return generateParamTable(
        reactors.Reactor, reactorParameters.defineReactorParameters()
